pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "andon-for-llm-agents"
version = "0.2.0"
description = "TPS-inspired ANDON safety hooks for LLM coding agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "AllNew LLC" }]
keywords = ["llm", "agents", "safety", "andon", "hooks", "claude-code"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = ["PyYAML>=5.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.9"]

[project.scripts]
andon = "hooks.andon_cli:main"

[project.urls]
Homepage = "https://github.com/allnew-llc/andon-for-llm-agents"
Issues = "https://github.com/allnew-llc/andon-for-llm-agents/issues"

[tool.setuptools]
packages = ["hooks", "hooks.vault", "hooks.vault.drivers"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["hooks"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["hooks"]
